Fixing ratings, fixing links: git push origin python

+++ src/main/java/skill_dev/models/entities/Grade.java
@@ -18,6 +18,6 @@ import javax.persistence.Table;
 public class Grade extends BaseEntity{
     @OneToOne
     Submission submission;
-    Float score;
+    Double score;
     String feedback;
 }
